Fix height building formatting
