+++ geoindicators/src/main/groovy/org/orbisgis/geoclimate/geoindicators/WorkflowGeoIndicators.groovy
@@ -1435,7 +1435,7 @@ Map estimateBuildingHeight(JdbcDataSource datasource, String zone, String buildi
             prefixName)
 
     def buildingTableName = "BUILDING_TABLE_WITH_RSU_AND_BLOCK_ID"
-    def nbBuildingEstimated
+    int nbBuildingEstimated =0
     def buildEstimatedHeight
     if (datasource.getTable(gatheredScales).isEmpty()) {
         info "No building height to estimate"
@@ -1460,29 +1460,29 @@ Map estimateBuildingHeight(JdbcDataSource datasource, String zone, String buildi
 
         nbBuildingEstimated = datasource.firstRow("select count(*) as count from $buildEstimatedHeight".toString()).count
 
-        datasource.createIndex(buildEstimatedHeight, "id_build")
+        //We must format only estimated buildings
+        //Apply format on the new abstract table
+        def epsg = datasource."$buildEstimatedHeight".srid;
+        def formatedBuildEstimatedHeight = formatEstimatedBuilding(datasource, buildEstimatedHeight, epsg)
 
-        def newEstimatedHeigthWithIndicators = "NEW_BUILDING_INDICATORS_${UUID.randomUUID().toString().replaceAll("-", "_")}"
+        datasource.createIndex(formatedBuildEstimatedHeight, "id_build")
+
+        buildingTableName = "INPUT_BUILDING_ESTIMATED_${UUID.randomUUID().toString().replaceAll("-", "_")}"
 
         //Use build table indicators
-        datasource.execute """DROP TABLE IF EXISTS $newEstimatedHeigthWithIndicators;
-                                               CREATE TABLE $newEstimatedHeigthWithIndicators as 
+        datasource.execute """DROP TABLE IF EXISTS $buildingTableName;
+                                               CREATE TABLE $buildingTableName as 
                                                 SELECT  a.THE_GEOM, a.ID_BUILD,a.ID_SOURCE,
                                             CASE WHEN b.HEIGHT_ROOF IS NULL THEN a.HEIGHT_WALL ELSE 0 END AS HEIGHT_WALL ,
                                                     COALESCE(b.HEIGHT_ROOF, a.HEIGHT_ROOF) AS HEIGHT_ROOF,
                                                     CASE WHEN b.HEIGHT_ROOF IS NULL THEN a.NB_LEV ELSE 0 END AS NB_LEV, a.TYPE,a.MAIN_USE, a.ZINDEX, a.ID_BLOCK, a.ID_RSU 
                                                 from $buildingIndicatorsForHeightEst
-                                            a LEFT JOIN $buildEstimatedHeight b on a.id_build=b.id_build""".toString()
-
-        //We must format only estimated buildings
-        //Apply format on the new abstract table
-        def epsg = datasource."$newEstimatedHeigthWithIndicators".srid;
-        buildingTableName = formatEstimatedBuilding(datasource, newEstimatedHeigthWithIndicators, epsg)
+                                            a LEFT JOIN $formatedBuildEstimatedHeight b on a.id_build=b.id_build""".toString()
 
 
         //Drop intermediate tables
         datasource.execute """DROP TABLE IF EXISTS $estimated_building_with_indicators,
-                                            $newEstimatedHeigthWithIndicators, $buildEstimatedHeight,
+                                            $formatedBuildEstimatedHeight, $buildEstimatedHeight,
                                             $gatheredScales""".toString()
     }
 
@@ -2164,7 +2164,6 @@ String formatEstimatedBuilding(JdbcDataSource datasource, String inputTableName,
         if (inputSpatialTable.rowCount > 0) {
             def columnNames = inputSpatialTable.columns
             queryMapper += " ${columnNames.join(",")} FROM $inputTableName"
-
             datasource.withBatch(100) { stmt ->
                 datasource.eachRow(queryMapper) { row ->
                     def heightRoof = row.height_roof
@@ -2191,6 +2190,63 @@ String formatEstimatedBuilding(JdbcDataSource datasource, String inputTableName,
     }
     info 'Re-formating building finishes'
     return outputTableName
+}
+
+
+/**
+ * Rule to guarantee the height wall, height roof and number of levels values
+ * @param heightWall value
+ * @param heightRoof value
+ * @param nbLevels value
+ * @param h_lev_min value
+ * @return a map with the new values
+ */
+static Map formatHeightsAndNbLevels(def heightWall, def heightRoof, def nbLevels, def h_lev_min,
+                                    def buildingType, def levelBuildingTypeMap) {
+    //Use the BDTopo values
+    if (heightWall != 0 && heightRoof != 0 && nbLevels != 0) {
+        return [heightWall: heightWall, heightRoof: heightRoof, nbLevels: nbLevels, estimated: false]
+    }
+    //Initialisation of heights and number of levels
+    // Update height_wall
+    boolean estimated = false
+    if (heightWall == 0) {
+        if (!heightRoof || heightRoof == 0) {
+            if (nbLevels == 0) {
+                nbLevels = levelBuildingTypeMap[buildingType]
+                if (!nbLevels) {
+                    nbLevels = 1
+                }
+                heightWall = h_lev_min * nbLevels
+                heightRoof = heightWall
+                estimated = true
+            } else {
+                heightWall = h_lev_min * nbLevels
+                heightRoof = heightWall
+            }
+        } else {
+            heightWall = heightRoof
+            nbLevels = Math.max(Math.floor(heightWall / h_lev_min), 1)
+        }
+    } else if (heightWall == heightRoof) {
+        if (nbLevels == 0) {
+            nbLevels = Math.max(Math.floor(heightWall / h_lev_min), 1)
+        }
+    }
+    // Control of heights and number of levels
+    // Check if height_roof is lower than height_wall. If yes, then correct height_roof
+    else if (heightWall > heightRoof) {
+        heightRoof = heightWall
+        if (nbLevels == 0) {
+            nbLevels = Math.max(Math.floor(heightWall / h_lev_min), 1)
+        }
+    } else if (heightRoof > heightWall) {
+        if (nbLevels == 0) {
+            nbLevels = Math.max(Math.floor(heightRoof / h_lev_min), 1)
+        }
+    }
+    return [heightWall: heightWall, heightRoof: heightRoof, nbLevels: nbLevels, estimated: estimated]
+
 }
